Apparently working.

+++ libft/libft.h
@@ -88,7 +88,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 
 char	**ft_split(char const *s, char c);
 
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long long n, int fd);
 
 void	ft_puthexanbr_fd(int n, int fd, char *digits);
